move all views into individual folders
