+++ source/app.js
@@ -22,10 +22,10 @@ import * as c from './views/components/colors'
 import {Title, LeftButton, RightButton, configureScene} from './views/components/navigation'
 
 import CalendarView from './views/calendar'
-import ContactsView from './views/contacts'
+import {ContactsView} from './views/contacts'
 import {DictionaryView, DictionaryDetailView} from './views/dictionary'
-import DirectoryView from './views/directory'
-import HomeView from './views/home'
+import {DirectoryView} from './views/directory'
+import {HomeView, EditHomeView} from './views/home'
 import MapView from './views/map'
 import StreamingView from './views/streaming'
 import {MenusView} from './views/menus'
@@ -41,7 +41,6 @@ import SISLoginView from './views/settings/login'
 import CreditsView from './views/settings/credits'
 import PrivacyView from './views/settings/privacy'
 import LegalView from './views/settings/legal'
-import EditHomeView from './views/editHome'
 import {StudentOrgsView, StudentOrgsDetailView}  from './views/student-orgs'
 import {FaqView} from './views/faqs'
