+++ source/views/contacts/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export {default as ContactsView} from './contacts'
