+++ source/views/directory/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export {default as DirectoryView} from './directory'
