+++ source/views/home/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export {default as HomeView} from './home'
+export {default as EditHomeView} from './edit'
